Update config files, controllers and services
Update config files to validate properly the roles and auth
Add swagger and notation in controllers
Update controllers and services to validate roles and map to dtos
Create a specifications file to allow filters in movies

+++ src/main/java/com/catalog/movies/services/MovieService.java
@@ -1,8 +1,10 @@
 package com.catalog.movies.services;
 
+import com.catalog.movies.dto.MovieResponseDTO;
 import com.catalog.movies.model.Movie;
 import com.catalog.movies.exception.ResourceNotFoundException;
 import com.catalog.movies.repository.MovieRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -12,31 +14,45 @@ import org.springframework.stereotype.Service;
 @Service
 public class MovieService {
 
+    private final MovieRepository movieRepository;
     @Autowired
-    private MovieRepository movieRepository;
+    private ModelMapper modelMapper;
 
-    public Movie createMovie(Movie movie) {
-        // Validate and save movie
-        return movieRepository.save(movie);
+    @Autowired
+    public MovieService(MovieRepository movieRepository) {
+        this.movieRepository = movieRepository;
     }
 
-    public Movie updateMovie(Long id, Movie updatedMovie) throws ResourceNotFoundException {
+    public MovieResponseDTO createMovie(Movie movie) {
+        return convertToDto(movieRepository.save(movie));
+    }
+
+
+    public MovieResponseDTO updateMovie(Long id, Movie updatedMovie) {
         Movie movie = movieRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Movie not found"));
-        // update fields and save
+                .orElseThrow(() -> new ResourceNotFoundException("Película no encontrada con id: " + id));
         movie.setName(updatedMovie.getName());
         movie.setSynopsis(updatedMovie.getSynopsis());
         movie.setReleaseYear(updatedMovie.getReleaseYear());
         movie.setCategories(updatedMovie.getCategories());
         movie.setPosterUrl(updatedMovie.getPosterUrl());
-        return movieRepository.save(movie);
+        // CreatedBy not updated
+        return convertToDto(movieRepository.save(movie));
     }
 
     public void deleteMovie(Long id) {
+        if (!movieRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Movie not found: id " + id);
+        }
         movieRepository.deleteById(id);
     }
 
-    public Page<Movie> getMovies(Specification<Movie> spec, Pageable pageable) {
-        return movieRepository.findAll(spec, pageable);
+    public Page<MovieResponseDTO> getMovies(Specification<Movie> spec, Pageable pageable) {
+        Page<Movie> movies = movieRepository.findAll(pageable);
+        return movies.map(this::convertToDto);
+    }
+
+    public MovieResponseDTO convertToDto(Movie movie) {
+        return modelMapper.map(movie, MovieResponseDTO.class);
     }
 }
